dirstream polling and multi-directory support
This commit makes one significant change and one addition to the dirstream client.

The change is that the client now uses directory polling instead of inotify to monitor the file directory. This adds a delay to the time taken for a file to get to the Strelka cluster, but fixes the race condition which resulted in incomplete files being sent to the cluster. It also decreases the complexity of the client, including avoiding performance issues caused by inotify (i.e. queue overflow). The delay incurred by this change is a minimal impact to high-volume production environments.

The change adds the ability to monitor multiple directories from one client.