Add ZeroMQ reconnection options
This commit makes two changes:

Ports have been redefined to “socket” ports. This results in no change in functionality, it’s simply a name change that should improve clarity in the configuration options.

Added options for setting ZeroMQ TCP reconnections on the task socket. This can help reduce network noise from the workers when the connection between them and the broker is disrupted.