Add Gatekeeper (#77)
* Init Gatekeeper component

“gatekeeper” is a system component (Redis database) that acts as a temporary event cache — this is useful for large scale deployments where the same file may be observed millions and millions of times per day.

Unlike the coordinator, gatekeeper is not compatible with Envoy due to the requirement of using transaction commands.

The following is still in the TODO for this component:
- clients have no way to bypass gatekeeper (requires change to protobuf, client apps/configs)
- gatekeeper should be an optional component

* Updates to frontend

This commit contains work in progress for bringing gatekeeper support to the frontend; however, the code is fully functional and working as designed.

* Add Gatekeeper Bypass

This commit adds a client option to bypass the gatekeeper. This addition affects filestream, fileshot, and the strelka protobufs. frontend was also refactored for readability and consistency.

* Added gatekeeper, removed cache

This commit adds detail about gatekeeper to the README and removes references to the cache (which is now part of coordinator).