handle missing 'MZ' PE file header
The PE parsing library we use expects there to be a 0x4D5A file header.
If the "DOS magic header" is missing then the libary fails out. This
change captures that exception and sets an event flag to indicate an
error occurred. Users who hit this error should review the scanned files
and their yara rules for tasting files.